two threads are working with mutex
