+++ app/prj.conf
@@ -7,17 +7,9 @@ CONFIG_CLOCK_CONTROL_NRF_K32SRC_RC=y
 # 500PPM enabled by default if RC selected
 #CONFIG_CLOCK_CONTROL_NRF_K32SRC_500PPM=y
 
-CONFIG_LOG=y
+#CONFIG_LOG=y
 
 #CONFIG_GPIO = y
-
-# Config Power Management
-# CONFIG_PM_CPU_OPS=y
-# CONFIG_PM_DEVICE=y
-# CONFIG_PM=y
-# CONFIG_PM_DEVICE_RUNTIME=y
-# CONFIG_PM_STATS=y
-# CONFIG_STATS=y
 
 # Enable System reboot functionality
 CONFIG_REBOOT=y
@@ -26,34 +18,26 @@ CONFIG_REBOOT=y
 # CONFIG_SERIAL=y
 # CONFIG_CONSOLE=y
 # CONFIG_UART_CONSOLE=y
+
 CONFIG_DEBUG=y
-#16384
-CONFIG_MAIN_STACK_SIZE=8192
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=8192
+#16384 8192
+CONFIG_MAIN_STACK_SIZE=2048
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
 
 #CONFIG_SENSOR=y
-CONFIG_BLINK=y
+#CONFIG_BLINK=y
 CONFIG_I2C=y
 CONFIG_SENSOR=y
-#CONFIG_BMI270=y
-#CONFIG_BMI05X=y
 CONFIG_BNO055=y
-#CONFIG_BNO055_BUS_I2C = y
-#CONFIG_BNO055_BUS_SPI = n
 
 CONFIG_SPI=y
 CONFIG_SPI_SLAVE=n
 
-#CONFIG_SOC_NRF52832_ALLOW_SPIM_DESPITE_PAN_58=y
 CONFIG_SPI_NRFX=y
-#CONFIG_SPI_NRFX_SPIM = y
-#CONFIG_SPI_NRFX_RAM_BUFFER_SIZE=32
 
-#CONFIG_ST7735R=y
 CONFIG_ST7735S=y
 CONFIG_DISPLAY=y
 
-#CONFIG_APP_LINK_WITH_LVGL=y
 CONFIG_LV_CONF_MINIMAL=y
 CONFIG_LVGL=y
 #CONFIG_LV_Z_BITS_PER_PIXEL=16
@@ -62,7 +46,7 @@ CONFIG_LV_MEM_CUSTOM=y
 #2048 default 8192 16384
 CONFIG_LV_Z_MEM_POOL_SIZE=2048
 
-CONFIG_LV_USE_LOG=y
+#CONFIG_LV_USE_LOG=y
 CONFIG_LV_USE_LABEL=y
 CONFIG_LV_USE_LINE=y
 #CONFIG_LV_FONT_MONTSERRAT_24=y
